+++ packages/calendar/src/views/month-grid/components/month-grid-wrapper.tsx
@@ -1,16 +1,18 @@
 import { PreactViewComponent } from '@schedule-x/shared/src/types/calendar/preact-view-component'
-import { useEffect, useState } from 'preact/hooks'
+import { useState } from 'preact/hooks'
 import { createMonth } from '../utils/stateless/create-month'
 import { Month } from '../types/month'
 import MonthGridWeek from './month-grid-week'
 import { AppContext } from '../../../utils/stateful/app-context'
 import { positionInMonth } from '../utils/stateless/position-in-month'
 import { sortEventsByStartAndEndWithoutConsideringTime } from '../../../utils/stateless/events/sort-by-start-date'
+import { filterByRange } from '../../../utils/stateless/events/filter-by-range'
+import { useSignalEffect } from '@preact/signals'
 
 export const MonthGridWrapper: PreactViewComponent = ({ $app, id }) => {
   const [month, setMonth] = useState<Month>([])
 
-  useEffect(() => {
+  useSignalEffect(() => {
     $app.calendarEvents.list.value.forEach((event) => {
       event._eventFragments = {}
     })
@@ -18,6 +20,17 @@ export const MonthGridWrapper: PreactViewComponent = ({ $app, id }) => {
       $app.datePickerState.selectedDate.value,
       $app.timeUnitsImpl
     )
+    newMonth.forEach((week) => {
+      week.forEach((day) => {
+        day.backgroundEvents = filterByRange(
+          $app.calendarEvents.backgroundEvents.value,
+          {
+            start: day.date,
+            end: day.date,
+          }
+        )
+      })
+    })
     const filteredEvents = $app.calendarEvents.filterPredicate.value
       ? $app.calendarEvents.list.value.filter(
           $app.calendarEvents.filterPredicate.value
@@ -29,12 +42,7 @@ export const MonthGridWrapper: PreactViewComponent = ({ $app, id }) => {
         filteredEvents.sort(sortEventsByStartAndEndWithoutConsideringTime)
       )
     )
-  }, [
-    $app.calendarState.range.value?.start,
-    $app.calendarState.range.value?.end,
-    $app.calendarEvents.list.value,
-    $app.calendarEvents.filterPredicate.value,
-  ])
+  })
 
   return (
     <AppContext.Provider value={$app}>