+++ packages/translations/src/locales/it-IT/calendar.ts
@@ -5,6 +5,7 @@ export const calendarItIT: CalendarTranslations = {
   Month: 'Mese',
   Week: 'Settimana',
   Day: 'Giorno',
+  'Select View': 'Seleziona la vista',
   events: 'eventi',
   event: 'evento',
   'No events': 'Nessun evento',