+++ packages/translations/src/locales/sk-SK/date-picker.ts
@@ -6,5 +6,4 @@ export const datePickerSkSK: DatePickerTranslations = {
   'Next month': 'Ďalší mesiac',
   'Previous month': 'Predchádzajúci mesiac',
   'Choose Date': 'Vyberte dátum',
-  'Select View': 'Vyberte zobrazenie',
 }