+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/schedule-x/schedule-x/compare/v2.4.3...v2.5.0) (2024-11-09)
+
+### Bug Fixes
+
+- reactivity errors from using preact useEffect for signals ([#770](https://github.com/schedule-x/schedule-x/issues/770)) ([79b3199](https://github.com/schedule-x/schedule-x/commit/79b3199fa1d3be45e117ad3be5f922bdb7621875))
+
+### Features
+
+- background events ([#646](https://github.com/schedule-x/schedule-x/issues/646)) ([489f396](https://github.com/schedule-x/schedule-x/commit/489f396302c9f9d745dfc75128c3aa0e58ac80da))
+
+## [2.4.3](https://github.com/schedule-x/schedule-x/compare/v2.4.2...v2.4.3) (2024-11-07)
+
+### Bug Fixes
+
+- missing \_options after dnd ([#761](https://github.com/schedule-x/schedule-x/issues/761)) ([ebe80a6](https://github.com/schedule-x/schedule-x/commit/ebe80a6ce3de5f834ddea6e35090aac661b0ccc2))
+
+## [2.4.2](https://github.com/schedule-x/schedule-x/compare/v2.4.1...v2.4.2) (2024-11-06)
+
+### Bug Fixes
+
+- **ical:** infer all-day events ([#724](https://github.com/schedule-x/schedule-x/issues/724)) ([af24a89](https://github.com/schedule-x/schedule-x/commit/af24a89cca2b0b122bf8f323f98eb217b48845ca))
+- remove node step in build workflow ([#754](https://github.com/schedule-x/schedule-x/issues/754)) ([4cbf89a](https://github.com/schedule-x/schedule-x/commit/4cbf89afbab9b18050647ab03676e80a5b02a80f))
+- update Dutch translations ([#745](https://github.com/schedule-x/schedule-x/issues/745)) ([151fa56](https://github.com/schedule-x/schedule-x/commit/151fa569a0cd53f0a1441f75ba41c0fe06e315e3))
+
 ## [2.4.1](https://github.com/schedule-x/schedule-x/compare/v2.4.0...v2.4.1) (2024-10-10)
 
 ### Bug Fixes